src_tauri::commands

Module file_system_operation_commands

Re-exports

pub use __cmd__open_file;

pub use __cmd__open_in_default_app;

pub use __cmd__open_directory;

pub use __cmd__create_file;

pub use __cmd__create_directory;

pub use __cmd__rename;

pub use __cmd__move_to_trash;

pub use __cmd__copy_file_or_dir;

pub use __cmd__zip;

pub use __cmd__unzip;

Functions

copy_file_or_dir

Copies a file or directory from the source path to the destination path. This function does not create any parent directories. It will overwrite the destination if it already exists. If the source is a directory, it will recursively copy all files and subdirectories.

create_directory

Creates a directory at the given absolute path. Returns a string if there was an error. This function does not create any parent directories.

create_file

Creates a file at the given absolute path. Returns a string if there was an error. This function does not create any parent directories.

move_to_trash

Deletes a file at the given path. Returns a string if there was an error. This function moves the file to the trash instead of deleting it permanently.

open_directory

Opens a directory at the given path and returns its contents as a json string.

open_file

Opens a file at the given path and returns its contents as a string. Should only be used for text files.

open_in_default_app

rename

Renames a file or directory at the given path.

unzip

Extracts zip files to specified destinations. If extracting a single zip file without a specified destination, extracts to a directory with the same name as the zip file.

zip

Zips files and directories to a destination zip file. If only one source path is provided and no destination is specified, creates a zip file with the same name. For multiple source paths, the destination path must be specified.