src_tauri::commands

Module hash_commands

Re-exports

pub use __cmd__gen_hash_and_return_string;

pub use __cmd__gen_hash_and_save_to_file;

pub use __cmd__compare_file_or_dir_with_hash;

Enums

ChecksumMethod

HashError

Functions

calculate_crc32 🔒

calculate_hash 🔒

calculate_md5 🔒

calculate_sha256 🔒

calculate_sha384 🔒

calculate_sha512 🔒

compare_file_or_dir_with_hash

Compares a file’s generated hash with a provided hash value. The hash algorithm used is determined by the application settings.

compare_file_or_dir_with_hash_impl

gen_hash_and_return_string

Generates a hash for the given file and returns it as a string. The hash algorithm used is determined by the application settings (MD5, SHA256, SHA384, SHA512, or CRC32).

gen_hash_and_return_string_impl

gen_hash_and_save_to_file

Generates a hash for the given file and saves it to a specified output file. The hash algorithm used is determined by the application settings.

gen_hash_and_save_to_file_impl

get_checksum_method 🔒

read_file 🔒