src_tauri::models

Function get_access_permission_string

pub fn get_access_permission_string(
    permissions: Permissions,
    is_directory: bool,
) -> String

This function converts the access permissions of a file or directory into a human-readable string. It takes into account the platform (Windows or Unix) and formats the permissions accordingly.

Parameters

permissions: The permissions of the file or directory.

is_directory: A boolean indicating whether the entry is a directory or not.

Returns

A string representing the access permissions in a human-readable format.

Example

use crate::commands::fs_dir_loader_commands::get_access_permission_string;
use std::fs::Permissions;
use std::os::unix::fs::PermissionsExt;

fn main() {
  let permissions = Permissions::from_mode(0o755);
  let is_directory = true;
  let permission_string = get_access_permission_string(permissions, is_directory);
  println!("Access permissions: {}", permission_string);
}